jordan: model interface
Added model interface with sample model (not great, but decent). The
test model is in /models, and all test images (for the sake of
consistency and .gitignore being able to properly ignore things) should
be in /tests. If /tests does not exist as a directory, create one.

The only dependency required for the model is torch and torchvision, no
CUDA/GPU required. Numpy and OpenCV should come with torch and
torchvision when it is pip-installed.